Updated details of proposed APIs
Dropped "from:" in symmetricDifference APIs.  It's too close to "form"
and not pulling its weight, especially in the nonmutating case where it
looks out-of-place.

This was considered:

  x.formXXX(with: y)

but "with" tends to make the operation sound as active in y as in x,
when in this case it is truly more active in x.  This is a marginal
difference, but one that I noticed when trying out both names.